requirements.txt
google-adk
google-cloud-aiplatform
openai
litellm
pandas
matplotlib
seaborn
